refactor: move transactions tests to integration
