+++ .evergreen/run-serverless-tests.sh
@@ -18,6 +18,7 @@ npx mocha --file test/tools/runner/index.js \
   test/integration/retryable-writes/retryable_writes.spec.test.js \
   test/integration/sessions/sessions.spec.test.js \
   test/integration/sessions/sessions.test.js \
-  test/functional/transactions.test.js \
+  test/integration/transactions/transactions.spec.test.js \
+  test/integration/transactions/transactions.test.js \
   test/integration/versioned-api/versioned_api.spec.test.js \
   test/integration/load-balancers/load_balancers.spec.test.js
